Refactored type placement
